enable fil-actor
