+++ ipc/sdk/Cargo.toml
@@ -9,9 +9,7 @@ license-file.workspace = true
 
 [dependencies]
 anyhow = "1.0.56"
-fil_actors_runtime = { git = "https://github.com/consensus-shipyard/fvm-utils", optional = true, features = [
-  "fil-actor",
-] }
+fil_actors_runtime = { workspace = true, optional = true }
 fnv = "1.0.7"
 fvm_ipld_blockstore = { workspace = true }
 fvm_ipld_encoding = { workspace = true }
@@ -29,7 +27,7 @@ thiserror = { workspace = true }
 
 [dev-dependencies]
 serde_json = "1.0.95"
-
+fil_actors_runtime = { workspace = true }
 
 [features]
 default = ["fil-actor"]
