+++ ipc/sdk/src/lib.rs
@@ -12,6 +12,8 @@ pub mod error;
 pub mod gateway;
 pub mod subnet;
 pub mod subnet_id;
+#[cfg(feature = "fil-actor")]
+mod runtime;
 
 /// Encodes the a ChainEpoch as a varInt for its use
 /// as a key of a HAMT. This serialization has been
